Lists added
They can't quite store tasks yet, nor can they themselves be saved, but when you hit create list in the dialog, it will give you a nice toolbar with the list name on it. (Xmls for tasks and the list have been made, just not wired up yet)

Also removed the list create activity, as this worked so well.