Task-based local saves complete
You can now successfully save tasks, with their due dates, and current status, names, and description. When loading a list, each task is checked for a calendar, if a calendar exists, it runs the timer check to update before being put on screen. If a timer reaches it's due date, it has it's calendar and timer negated to prevent confusion

More task-based settings to come from saving, but for now this works perfectly fine and I'm quite happy with it