Categories support
Categories is now fully supported and operational in the system. Each list can be assigned a category from the list settings button (list color relegated to a button within that dialog)

Each new category tag generates a new tag to search by on the main menu. When you select a tag from the dropdown menu, only lists with that specific tag will show up until the page is reloaded, or a new option is picked.

All resets this and shows all lists. Search functionality can be used with tags to further narrow down visible lists.