Task gui overhaul
Updating task list gui to implement the card system I generated for the list of lists.

Task state (checked off, failed, or empty) is shown in the background of the card with a low opacity png file.

Cards have 1 line for title, 2 lines for date (cuts off time if it doesn't on smaller devices) and 2 lines for description.

Tapping a task card generates the response of tapping the check boxes of the update prior. Tapping and holding for a second pops up the edit dialog.

I also finally got around to removing the old debug code for background downloading.